delivery/pyproject.toml
[project]
name = "delivery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.5.8",
    "fastapi[standard]>=0.123.5",
    "psycopg[binary]>=3.3.1",
    "pyfaulttolerance>=0.1.3",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.27",
]
